+++ src/fragment-components/index.ts
@@ -18,7 +18,7 @@ import * as row from './a-row';
 import * as tag from './a-tag';
 import * as searchinput from './a-searchinput';
 import * as text from './a-text';
-import * as textarea from './a-textarea';
+import * as textarea from './a-text-area';
 import * as textinput from './a-textinput';
 import * as link from './a-link';
 import * as loading from './a-loading';
@@ -56,7 +56,7 @@ export { ARow, ARowSettingsUI } from './a-row';
 export { ASearchInput, ASearchInputSettingsUI, ASearchInputCodeUI } from './a-searchinput';
 export { ATag, ATagSettingsUI } from './a-tag';
 export { AText, ATextSettingsUI } from './a-text';
-export { ATextArea, ATextAreaSettingsUI, ATextAreaCodeUI } from './a-textarea';
+export { ATextArea, ATextAreaSettingsUI, ATextAreaCodeUI } from './a-text-area';
 export { ATextInput, ATextInputSettingsUI, ATextInputCodeUI } from './a-textinput';
 export { AOverflowMenu, AOverflowMenuCodeUI, AOverflowMenuSettingsUI } from './a-overflow-menu';
 export { ARadio, ARadioSettingsUI, ARadioCodeUI } from './a-radio';
